Use query.name as the name of the entry
This is useful if different output builds are used. E.g. an es5 and es6 build.
Without this option, one build overrides the results of the other one.
Having the option to pass the name, combined with a configured output options
for the worker loader, different output files can be targeted (without
to have using hashes).