feat(options): add options.runtimeImportFunctionName
This feature allows a browser to determine the path from where to import
the worker script file from.

I decided to not use publicPath because I concluded that its value is best to have it resolved during build time by webpack.
Instead, runtimeImportFunctionName is meant to be used during runtime.

Addresses issue #161